Fix resolved hit HP loss type conversion

+++ Assets/Scripts/Game/Battle/BattleEffectLifecycle.cs
@@ -10,7 +10,7 @@ namespace GuildAdventure.Game.Battle
         public string actionId,sourceId,targetId;
         public int hitIndex;
         public string judgement;
-        public int actualHpLoss;
+        public double actualHpLoss;
     }
 
     public sealed class BattleTriggerDispatch
